Add :visited + ect. to links.css
Hey saw the issue about links not changing color after visiting, so I thought I would propose this. The visited color is a muted blue and the hover color is a golden yellow. You can change the colors if you wish, but I feel these would be a good base.